deployment/requirements.txt
# CDK dependencies
aws-cdk-lib>=2.166.0
aws-cdk.aws-lambda-python-alpha==2.166.0-alpha.0
constructs>=10.0.0

# AWS SDK and utilities
#boto3>=1.26.0
botocore>=1.29.0

# Development dependencies
pytest>=7.0.0
black>=22.0.0
pylint>=2.15.0
requests_aws4auth
opensearch-py
../boto3-1.35.72-py3-none-any.whl
../botocore-1.35.72-py3-none-any.whl

# streamlit>=1.40.0
